Moved parcels to solar_setbacks file
